django messages

+++ base/views.py
@@ -127,11 +127,27 @@ def updateRoom(request,pk):
 @login_required(login_url='login')
 def deleteRoom(request,pk):
     room= Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj': room})
 
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!')
+
+    if request.method == "POST":
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': message})
 
 def sztoki(request):
     return render(request, 'base/sztoki.html')
